requirements.txt
PyQt6
matplotlib
Pillow
